backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis & Celery
redis==5.0.1
celery==5.3.6
flower==2.0.1

# OCR & Image Processing
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0
python-magic==0.4.27

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client for CRM
httpx==0.26.0
aiohttp==3.9.1

# Utilities
phonenumbers==8.13.27
python-dateutil==2.8.2
# spacy==3.7.2  # Optional - pure regex works great without it

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
